backend/python-fastapi/requirements.txt
fastapi[standard]
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
mysql-connector-python
